+++ src/parser/OpCode0502.cpp
@@ -101,11 +101,11 @@ namespace OpenLogReplicator {
                                          ctx->read32(redoLogRecord->data() + fieldPos + 28));
 
             *ctx->dumpStream << "ktudh redo:" <<
-                             " slt: 0x" << std::setfill('0') << std::setw(4) << std::hex << static_cast<uint64_t>(redoLogRecord->xid.slt()) <<
+                             " slt: 0x" << std::setfill('0') << std::setw(4) << std::hex << static_cast<uint>(redoLogRecord->xid.slt()) <<
                              " sqn: 0x" << std::setfill('0') << std::setw(8) << std::hex << redoLogRecord->xid.sqn() <<
                              " flg: 0x" << std::setfill('0') << std::setw(4) << redoLogRecord->flg <<
                              " siz: " << std::dec << siz <<
-                             " fbi: " << std::dec << static_cast<uint64_t>(fbi) << '\n';
+                             " fbi: " << std::dec << static_cast<uint>(fbi) << '\n';
             /*if (ctx->version < RedoLogRecord::REDO_VERSION_12_1 || redoLogRecord->conId == 0)
                 *ctx->dumpStream << "           " <<
                         " uba: " << PRINTUBA(uba) << "   " <<
